Add simple test that the module is effective.
We can test that the module isn't completely broken by trying to load
it. That's one simple test.

Then we can try to run a small BASIC program that reports an error and
see whether we exit nicely. If we do, then we passed - but we can also
check whether it exited with the correct code and in the right amount
of time.